5/9 Saving SheeData through streaming: Move methods for saving the worksheet from XLWorkbook to WorksheetPartWriter. Methods are just copied, no change in code (except PopulatedRichTextElements and PopulateAutoFilter that have change access modifer)
